BackgroundContainer 생성
컴포넌트 계층 구조 반영
